e2e_samples/unstructured_data/src/requirements.txt
azure-storage-blob>=12.5.0,<12.20.0
python-dotenv>=1.0.1,<2.0.0
azure-ai-formrecognizer>=3.3.3,<4.0.0
PyMuPDF>=1.24.11,<2.0.0 # fitz
azure-ai-inference>=1.0.0b8,<2.0.0
azure-monitor-opentelemetry>=1.6.2,<2.0.0
opentelemetry-sdk>=1.27.0,<2.0.0
opentelemetry-instrumentation-openai>=0.33.0,<1.0.0
pandas>=2.2.3,<3.0.0
jinja2>=3.1.4,<4.0.0
promptflow-tracing>=1.16.2,<2.0.0
pyodbc>=5.2.0,<6.0.0
azure-identity>=1.19.0,<2.0.0
rouge-score>=0.1.2,<1.0.0
azure-ai-ml>=1.23.1,<2.0.0
azureml-core>=1.32.0,<2.0.0
mlflow>=2.19.0,<3.0.0
azureml-mlflow>=1.59.0.post1,<2.0.0
azure-ai-evaluation>=1.1.0,<2.0.0
